backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic==1.10.13
python-multipart
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose
email-validator
pymysql
pyyaml 
ua-parser 
user-agents
httpx
fastapi-mail==1.3.1
firebase
firebase-admin
qrcode
Pillow
stripe